function/save_game/requirements.txt
functions-framework==3.*
fastapi
pymongo
python-dotenv